+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## []
 
 - Added damage and attack speed modifier for Hammer, Saw, File and Sharp Flint
+- Cake and Cookie now use Flour in recipe
 
 ## [0.4.5]